refactor : 직접 작성한 테스트 코드 이동
원래 존재하던 테스트 코드와 직접 작성한 테스트 코드를 서로 다른 파일로 분리하였습니다.
